VTOL Weathervane functionaliy
vtol_att_control

Orients the VTOL aircraft with the nose facing the wind.

This is done by using the (nonzero) roll setpoint necessary to stay on a fixed position to determine in which direction to turn in order to turn nose to wind. When nose is against wind only pitch/pusher will be necessary for staying in one position.

Signed-off-by: Ivo Drescher <ivo.drescher@gmail.com>